Feature get all device and get device by id

+++ Domain/Entities/DeviceDailyAvailability.cs
@@ -4,6 +4,6 @@ public class DeviceDailyAvailability
 {
     public DateOnly DateTime { get; set; }
     public int DeviceID { get; set; }
-    public int TotalAvailableMinutes { get; set; }
+    public double TotalAvailableMinutes { get; set; }
     public Device Device { get; set; }
 }
